Started implementing Azure Sync services / First draft of Income - Expenses section

+++ FinalApp/Models/UserIncome.cs
@@ -6,7 +6,7 @@ namespace FinalApp.Models {
     using J = Newtonsoft.Json.JsonPropertyAttribute;
 
     public partial class UserIncome : BindableObject {
-        [J("amount")] public long Amount { get; set; }
+        [J("amount")] public double Amount { get; set; }
         [J("description")] public string Description { get; set; }
         [J("incomeDate")] public DateTimeOffset IncomeDate { get; set; }
         [J("categoryId")] public long CategoryId { get; set; }
